+++ osu.Game/Online/Chat/ChannelManager.cs
@@ -25,6 +25,7 @@ namespace osu.Game.Online.Chat
         /// </summary>
         private readonly string[] defaultChannels =
         {
+            @"#chinese",
             @"#lazer",
             @"#osu",
             @"#lobby"
@@ -123,7 +124,7 @@ namespace osu.Game.Online.Chat
             {
                 if (!api.IsLoggedIn)
                 {
-                    target.AddNewMessages(new ErrorMessage("Please sign in to participate in chat!"));
+                    target.AddNewMessages(new ErrorMessage("请在登录后进行发言!"));
                     return;
                 }
 
@@ -152,7 +153,7 @@ namespace osu.Game.Online.Chat
 
                     createNewPrivateMessageRequest.Failure += exception =>
                     {
-                        Logger.Error(exception, "Posting message failed.");
+                        Logger.Error(exception, "发送消息失败oAo");
                         target.ReplaceMessage(message, null);
                         dequeueAndRun();
                     };
@@ -171,7 +172,7 @@ namespace osu.Game.Online.Chat
 
                 req.Failure += exception =>
                 {
-                    Logger.Error(exception, "Posting message failed.");
+                    Logger.Error(exception, "发送消息失败oAo");
                     target.ReplaceMessage(message, null);
                     dequeueAndRun();
                 };
@@ -206,7 +207,7 @@ namespace osu.Game.Online.Chat
                 case "me":
                     if (string.IsNullOrWhiteSpace(content))
                     {
-                        target.AddNewMessages(new ErrorMessage("Usage: /me [action]"));
+                        target.AddNewMessages(new ErrorMessage("用法: /me [动作]"));
                         break;
                     }
 
@@ -216,7 +217,7 @@ namespace osu.Game.Online.Chat
                 case "join":
                     if (string.IsNullOrWhiteSpace(content))
                     {
-                        target.AddNewMessages(new ErrorMessage("Usage: /join [channel]"));
+                        target.AddNewMessages(new ErrorMessage("用法: /join [频道id]"));
                         break;
                     }
 
@@ -224,7 +225,7 @@ namespace osu.Game.Online.Chat
 
                     if (channel == null)
                     {
-                        target.AddNewMessages(new ErrorMessage($"Channel '{content}' not found."));
+                        target.AddNewMessages(new ErrorMessage($"频道 '{content}' 未找到."));
                         break;
                     }
 
@@ -233,11 +234,11 @@ namespace osu.Game.Online.Chat
                     break;
 
                 case "help":
-                    target.AddNewMessages(new InfoMessage("Supported commands: /help, /me [action], /join [channel]"));
+                    target.AddNewMessages(new InfoMessage("当前版本支持的指令: /help, /me [动作], /join [频道id]"));
                     break;
 
                 default:
-                    target.AddNewMessages(new ErrorMessage($@"""/{command}"" is not supported! For a list of supported commands see /help"));
+                    target.AddNewMessages(new ErrorMessage($@"""/{command}"" 目前还没有加入! 请输入 /help 来查看当前版本支持的指令."));
                     break;
             }
         }
@@ -269,7 +270,7 @@ namespace osu.Game.Online.Chat
             };
             req.Failure += error =>
             {
-                Logger.Error(error, "Fetching channel list failed");
+                Logger.Error(error, "获取频道列表失败");
                 initializeChannels();
             };
 
@@ -458,7 +459,7 @@ namespace osu.Game.Online.Chat
     public class ChannelNotFoundException : Exception
     {
         public ChannelNotFoundException(string channelName)
-            : base($"A channel with the name {channelName} could not be found.")
+            : base($"频道 {channelName} 未找到")
         {
         }
     }