+++ osu.Game/Online/Chat/StandAloneChatDisplay.cs
@@ -58,7 +58,7 @@ namespace osu.Game.Online.Chat
                 {
                     RelativeSizeAxes = Axes.X,
                     Height = textbox_height,
-                    PlaceholderText = "在此输入你的信息",
+                    PlaceholderText = "在此输入你要发送的信息",
                     OnCommit = postMessage,
                     ReleaseFocusOnCommit = false,
                     HoldFocus = true,