+++ osu.Game/Overlays/Profile/Header/BottomHeaderContainer.cs
@@ -53,12 +53,12 @@ namespace osu.Game.Overlays.Profile.Header
                     Spacing = new Vector2(0, 10),
                     Children = new Drawable[]
                     {
-                        topLinkContainer = new LinkFlowContainer(text => text.Font = text.Font.With(size: 14))
+                        topLinkContainer = new LinkFlowContainer(text => text.Font = text.Font.With(size: 18))
                         {
                             RelativeSizeAxes = Axes.X,
                             AutoSizeAxes = Axes.Y,
                         },
-                        bottomLinkContainer = new LinkFlowContainer(text => text.Font = text.Font.With(size: 14))
+                        bottomLinkContainer = new LinkFlowContainer(text => text.Font = text.Font.With(size: 18))
                         {
                             RelativeSizeAxes = Axes.X,
                             AutoSizeAxes = Axes.Y,