+++ osu.Game/Overlays/Profile/Header/Components/ExpandDetailsButton.cs
@@ -15,7 +15,7 @@ namespace osu.Game.Overlays.Profile.Header.Components
     {
         public readonly BindableBool DetailsVisible = new BindableBool();
 
-        public override string TooltipText => DetailsVisible.Value ? "collapse" : "expand";
+        public override string TooltipText => DetailsVisible.Value ? "收起" : "展开";
 
         private SpriteIcon icon;