+++ osu.Game/Overlays/Profile/Sections/Ranks/DrawablePerformanceScore.cs
@@ -39,7 +39,7 @@ namespace osu.Game.Overlays.Profile.Sections.Ranks
                     Anchor = Anchor.TopRight,
                     Origin = Anchor.TopRight,
                     Colour = colour.GrayA,
-                    Font = OsuFont.GetFont(size: 11, weight: FontWeight.Regular, italics: true)
+                    Font = OsuFont.GetFont(size: 15, weight: FontWeight.Regular, italics: true)
                 });
             }
         }